event-management-api/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
psycopg==3.1.18
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
